Factor symbols to an instance
