+++ browser/src/Services/Language/index.ts
@@ -12,4 +12,3 @@ export * from "./LanguageConfiguration"
 export * from "./LanguageEditorIntegration"
 export * from "./LanguageManager"
 export * from "./SignatureHelp"
-export * from "./Symbols"
